Move bup test into bup.rs
